Actualziacion ORM

+++ src/main/java/com/bddp1/dao/ProductoDAO.java
@@ -4,5 +4,5 @@ import com.bddp1.model.ProductoVendido;
 
 public interface ProductoDAO {
     public ProductoVendido getProdMoreRequested(String territory);
-    public boolean updateStock(Long category);
+    public boolean updateStock(int category);
 }
