+++ src/main/java/com/bddp1/dao/VentaDAO.java
@@ -4,5 +4,5 @@ import com.bddp1.model.Venta;
 import java.util.List;
 
 public interface VentaDAO {
-    public List<Venta> getTotalVentasProducto(Long category);
+    public List<Venta> getTotalVentasProducto(int category);
 }
